国・都道府県・川名でデータを取得
